test: move caching tests to their own file
