+++ ibis/backends/tests/test_client.py
@@ -1336,164 +1336,6 @@ def test_create_table_timestamp(con, temp_table):
     assert result == schema
 
 
-@mark.notimpl(["datafusion", "flink", "impala", "trino", "druid"])
-@mark.never(
-    ["mssql"],
-    reason="mssql supports support temporary tables through naming conventions",
-)
-@mark.notimpl(["exasol"], reason="Exasol does not support temporary tables")
-@pytest.mark.never(
-    ["risingwave"],
-    raises=com.UnsupportedOperationError,
-    reason="Feature is not yet implemented: CREATE TEMPORARY TABLE",
-)
-def test_persist_expression(backend, alltypes):
-    non_persisted_table = alltypes.mutate(
-        test_column=ibis.literal("calculation"), other_calc=ibis.literal("xyz")
-    )
-    persisted_table = non_persisted_table.cache()
-    backend.assert_frame_equal(
-        non_persisted_table.to_pandas(), persisted_table.to_pandas()
-    )
-
-
-@mark.notimpl(["datafusion", "flink", "impala", "trino", "druid"])
-@mark.never(
-    ["mssql"],
-    reason="mssql supports support temporary tables through naming conventions",
-)
-@mark.notimpl(["exasol"], reason="Exasol does not support temporary tables")
-@pytest.mark.never(
-    ["risingwave"],
-    raises=com.UnsupportedOperationError,
-    reason="Feature is not yet implemented: CREATE TEMPORARY TABLE",
-)
-def test_persist_expression_contextmanager(backend, con, alltypes):
-    non_cached_table = alltypes.mutate(
-        test_column=ibis.literal("calculation"), other_column=ibis.literal("big calc")
-    )
-    with non_cached_table.cache() as cached_table:
-        backend.assert_frame_equal(
-            non_cached_table.to_pandas(), cached_table.to_pandas()
-        )
-    assert non_cached_table.op() not in con._query_cache.cache
-
-
-@mark.notimpl(["datafusion", "flink", "impala", "trino", "druid"])
-@mark.never(
-    ["mssql"],
-    reason="mssql supports support temporary tables through naming conventions",
-)
-@mark.notimpl(["exasol"], reason="Exasol does not support temporary tables")
-@pytest.mark.never(
-    ["risingwave"],
-    raises=com.UnsupportedOperationError,
-    reason="Feature is not yet implemented: CREATE TEMPORARY TABLE",
-)
-def test_persist_expression_contextmanager_ref_count(backend, con, alltypes):
-    non_cached_table = alltypes.mutate(
-        test_column=ibis.literal("calculation"), other_column=ibis.literal("big calc 2")
-    )
-    op = non_cached_table.op()
-
-    with non_cached_table.cache() as cached_table:
-        assert op in con._query_cache.cache
-        backend.assert_frame_equal(
-            non_cached_table.to_pandas(), cached_table.to_pandas()
-        )
-
-    assert op not in con._query_cache.cache
-
-
-@mark.notimpl(["datafusion", "flink", "impala", "trino", "druid"])
-@mark.never(
-    ["mssql"],
-    reason="mssql supports support temporary tables through naming conventions",
-)
-@pytest.mark.never(
-    ["risingwave"],
-    raises=com.UnsupportedOperationError,
-    reason="Feature is not yet implemented: CREATE TEMPORARY TABLE",
-)
-@mark.notimpl(["exasol"], reason="Exasol does not support temporary tables")
-def test_persist_expression_multiple_refs(backend, con, alltypes):
-    non_cached_table = alltypes.mutate(
-        test_column=ibis.literal("calculation"), other_column=ibis.literal("big calc 2")
-    )
-    op = non_cached_table.op()
-    cached_table = non_cached_table.cache()
-
-    backend.assert_frame_equal(non_cached_table.to_pandas(), cached_table.to_pandas())
-
-    name = cached_table.op().name
-    nested_cached_table = non_cached_table.cache()
-
-    # cached tables are identical and reusing the same op
-    assert cached_table.op() is nested_cached_table.op()
-    # table is cached
-    assert op in con._query_cache.cache
-
-    # deleting the first reference, leaves table in cache
-    del nested_cached_table
-    assert op in con._query_cache.cache
-
-    # deleting the last reference, releases table from cache
-    del cached_table
-    assert op not in con._query_cache.cache
-
-    # assert that table has been dropped
-    assert name not in con.list_tables()
-
-
-@mark.notimpl(["datafusion", "flink", "impala", "trino", "druid"])
-@mark.never(
-    ["mssql"],
-    reason="mssql supports support temporary tables through naming conventions",
-)
-@mark.notimpl(["exasol"], reason="Exasol does not support temporary tables")
-@pytest.mark.never(
-    ["risingwave"],
-    raises=com.UnsupportedOperationError,
-    reason="Feature is not yet implemented: CREATE TEMPORARY TABLE",
-)
-def test_persist_expression_repeated_cache(alltypes, con):
-    non_cached_table = alltypes.mutate(
-        test_column=ibis.literal("calculation"), other_column=ibis.literal("big calc 2")
-    )
-    cached_table = non_cached_table.cache()
-    nested_cached_table = cached_table.cache()
-    name = cached_table.op().name
-
-    assert not nested_cached_table.to_pandas().empty
-
-    del nested_cached_table, cached_table
-
-    assert name not in con.list_tables()
-
-
-@mark.notimpl(["datafusion", "flink", "impala", "trino", "druid"])
-@mark.notimpl(["exasol"], reason="Exasol does not support temporary tables")
-@pytest.mark.never(
-    ["risingwave"],
-    raises=com.UnsupportedOperationError,
-    reason="Feature is not yet implemented: CREATE TEMPORARY TABLE",
-)
-def test_persist_expression_release(con, alltypes):
-    non_cached_table = alltypes.mutate(
-        test_column=ibis.literal("calculation"), other_column=ibis.literal("big calc 3")
-    )
-    cached_table = non_cached_table.cache()
-    cached_table.release()
-
-    assert non_cached_table.op() not in con._query_cache.cache
-
-    # a second release does not hurt
-    cached_table.release()
-
-    with pytest.raises(Exception, match=cached_table.op().name):
-        cached_table.execute()
-
-
 @contextlib.contextmanager
 def gen_test_name(con: BaseBackend):
     name = gen_name("test_table")
